add optional cloning of examples directory
a separate examples directory is maintained so that
large mesh files are not stored directly in goal's
git history (something git is not good at dealing with).

as a slight workaround, we sacrifice another repository
(git@github.com:bgranzow/goal-examples) to deal with large
revision changes due to the storage of mesh files.